% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_new.R
\name{createPeptidesCluster}
\alias{createPeptidesCluster}
\title{Create simulated cluster of peptides that share segments}
\usage{
createPeptidesCluster(segment_peptide_matrix, num_exchangeable)
}
\description{
Create simulated cluster of peptides that share segments
}
