% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{get_analytical_gradient_loss}
\alias{get_analytical_gradient_loss}
\title{Analytical gradient of PLS-GLS function}
\usage{
get_analytical_gradient_loss(
  observed_spectra,
  peptides_cluster,
  pept_seg_struct,
  num_parameters,
  time_0_data,
  undeuterated_dists,
  weights = NULL,
  theta = 1
)
}
\arguments{
\item{observed_spectra}{description}

\item{peptides_cluster}{description}

\item{pept_seg_struct}{description}

\item{num_parameters}{description}

\item{time_0_data}{description}

\item{undeuterated_dists}{description}

\item{weights}{description}

\item{theta}{description}
}
\description{
Analytical gradient of PLS-GLS function
}
