% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{getExchangeProbabilities}
\alias{getExchangeProbabilities}
\title{Get peptide-level probabilities from segment-level probabilities}
\usage{
getExchangeProbabilities(
  by_segment_probabilities,
  num_exchangeable,
  approximate_root = FALSE
)
}
\description{
Get peptide-level probabilities from segment-level probabilities
}
\keyword{internal}
