% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{getExpectedPeakHeights}
\alias{getExpectedPeakHeights}
\title{Calculate expected peak heights}
\usage{
getExpectedPeakHeights(
  spectrum,
  probabilities,
  undeuterated_dist,
  num_exchangeable
)
}
\description{
Calculate expected peak heights
}
\keyword{internal}
