% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{getExpectedProbabilitiesTable}
\alias{getExpectedProbabilitiesTable}
\title{Get expected exchange probabilities at the peptide level}
\usage{
getExpectedProbabilitiesTable(
  spectra,
  peptides_cluster,
  time_0_data,
  times,
  betas
)
}
\arguments{
\item{spectra}{xxx}

\item{peptides_cluster}{xxx}

\item{time_0_data}{xxx}

\item{times}{xxx}

\item{betas}{xxx}
}
\description{
Get expected exchange probabilities at the peptide level
}
