% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{getExpectedSegmentProbabilitiesTable}
\alias{getExpectedSegmentProbabilitiesTable}
\title{Expected exchange probabilities at the segment level}
\usage{
getExpectedSegmentProbabilitiesTable(
  spectra,
  peptides_cluster,
  time_0_data,
  times,
  betas
)
}
\arguments{
\item{spectra}{xxx}

\item{peptides_cluster}{xxx}

\item{time_0_data}{xxx}

\item{times}{xxx}

\item{betas}{xxx}
}
\description{
Expected exchange probabilities at the segment level
}
