% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{getExpectedSpectra}
\alias{getExpectedSpectra}
\title{Calculate expected isotopic peaks based on fixed values of beta parameters}
\usage{
getExpectedSpectra(spectra, peptides_cluster, times, betas)
}
\arguments{
\item{spectra}{data.table of observed spectra}

\item{peptides_cluster}{data.table with information about peptides cluster}

\item{times}{numeric vector of time points}

\item{beta}{numeric vector of fixed values of beta parameters}
}
\description{
Calculate expected isotopic peaks based on fixed values of beta parameters
}
\keyword{internal}
