% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{getOptimizationProblem}
\alias{getOptimizationProblem}
\title{Create optimization problem to estimate segment-specific exchange rates via OLS}
\usage{
getOptimizationProblem(
  observed_spectra,
  peptides_cluster,
  time_0_data,
  undeuterated_dists,
  weights = NULL,
  theta = 1
)
}
\arguments{
\item{observed_spectra}{data.table with observed isotopic distributions}

\item{peptides_cluster}{data.table with information about peptides cluster}

\item{time_0_data}{data.table with information about isotopic distributions of
undeuterated peptides}

\item{weights}{optional data.table with weights for each time point, peptide and isotopic peak}

\item{theta}{optional scalar variance parameter (see more in the Details section)}
}
\description{
Create optimization problem to estimate segment-specific exchange rates via OLS
}
