% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{getProbabilitiesFromBetas}
\alias{getProbabilitiesFromBetas}
\title{Convert beta parameters to probabilities}
\usage{
getProbabilitiesFromBetas(betas, time)
}
\description{
Convert beta parameters to probabilities
}
\keyword{internal}
