% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{getSpectra}
\alias{getSpectra}
\title{Simulate spectra for given sequences and time points}
\usage{
getSpectra(
  peptide_sequences,
  nums_exchangeable,
  time_points,
  lambdas,
  total_intensity = 1e+06,
  error_sd = 0.1
)
}
\arguments{
\item{peptide_sequences}{character vector of peptide sequences}

\item{nums_exchangeable}{numeric vector of exchangeable hydrogens counts for each peptide}

\item{time_points}{numeric vector of time points}

\item{lambdas}{numeric vector of parameters for atom-specific exchange rates}

\item{total_intensity}{total intensities in a spectrum}

\item{error_sd}{error for intensities (standard deviation of N(0, error_sd))}
}
\value{
data.table
}
\description{
Simulate spectra for given sequences and time points
}
