% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{makeSegmentProbabilities}
\alias{makeSegmentProbabilities}
\title{Get segment-level exchange probabilities from beta parameters}
\usage{
makeSegmentProbabilities(betas, times, num_parameters, present_segments)
}
\description{
Get segment-level exchange probabilities from beta parameters
}
\keyword{internal}
