% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_new.R
\name{simulateSegmentModel}
\alias{simulateSegmentModel}
\title{Simulate spectra from a model that relates links peptide-level spectra to segment-level exchange probabilities}
\usage{
simulateSegmentModel(
  peptides_cluster,
  time_points,
  beta_parameters,
  error_sd = 0.1,
  total = 1e+05
)
}
\description{
Simulate spectra from a model that relates links peptide-level spectra to segment-level exchange probabilities
}
