% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{fitIsoSegmentModel}
\alias{fitIsoSegmentModel}
\title{Fit isotopic distribution-based segment-level model}
\usage{
fitIsoSegmentModel(
  observed_spectra,
  peptides_cluster,
  time_0_data,
  undeuterated_dists,
  starting_point = NULL,
  method = "OLS",
  use_analytical_gradient = TRUE,
  max_iter = 100,
  tolerance = 0.01
)
}
\arguments{
\item{observed_spectra}{xx}

\item{peptides_cluster}{xx}

\item{time_0_data}{xx}

\item{undeuterated_dists}{xx}

\item{starting_point}{xx}

\item{method}{xx}

\item{use_analytical_gradient}{xx}
}
\description{
Fit isotopic distribution-based segment-level model
}
