% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{getOptimizationProblem}
\alias{getOptimizationProblem}
\title{Create optimization problem to estimate segment-specific exchange rates via OLS}
\usage{
getOptimizationProblem(
  observed_spectra,
  peptides_cluster,
  times,
  max_peaks = 10
)
}
\arguments{
\item{observed_spectra}{data.table with observed isotopic distributions}

\item{peptides_cluster}{data.table with information about peptides cluster}

\item{times}{numeric vector of time points}
}
\description{
Create optimization problem to estimate segment-specific exchange rates via OLS
}
